Add support for passing multiple arguments to VM
The eBPF spec supports passing arguments via registers R1 - R5[1].
However, uBPF only supported a single argument (via R1), so add support
for the remaining register.

On the test program, the user may now optionally specify the --mem
option up to 5 times, loading each file into R1 - R5.

A note on bounds checking: during execution, it is impossible to know to
which memory buffer the access refers to. Thus, we accept the access if
it falls within any of the memory buffers (i.e., it may overflow a
one buffer, but if it falls inside another, this is accepted).

Note that JIT is not yet supported. This will be added in a future
patch.

[1]: https://www.kernel.org/doc/Documentation/networking/filter.txt

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>